Refactor RBS
